move all game guis to the games tab
